rename,make

+++ src/main/java/com/project/yjshop/domain/image/ImageRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface ImageRepository extends CrudRepository<Image, Long> {
+public interface ImageRepository extends CrudRepository<Image, Integer> {
     Optional<Image> findByImagePath(String path);
 }
