+++ src/main/java/com/project/yjshop/service/user/UserService.java
@@ -12,7 +12,7 @@ import java.util.List;
 public interface UserService {
     UserProductResponse purchase(UserProductRequest userProductRequest, BindingResult bindingResult, PrincipalDetails principalDetails);
     UserProductResponse basket(UserProductRequest userProductRequest, BindingResult bindingResult, PrincipalDetails principalDetails);
-    UserProductResponse delBasket(Long boardId, PrincipalDetails principalDetails);
+    UserProductResponse delBasket(Integer boardId, PrincipalDetails principalDetails);
     List<Basket> myBasket(PrincipalDetails principalDetails);
 
 }
