+++ src/main/java/com/project/yjshop/web/api/UserController.java
@@ -45,7 +45,7 @@ public class UserController {
     }
 
     @DeleteMapping("/basket/{boardId}")
-    public UserProductResponse delBasket(@PathVariable Long boardId,
+    public UserProductResponse delBasket(@PathVariable Integer boardId,
                                          @AuthenticationPrincipal PrincipalDetails principalDetails) {
         return userService.delBasket(boardId, principalDetails);
     }
